BP-4262: Remove pathways and recipient apps
