+++ apps/issuer/admin.py
@@ -118,7 +118,7 @@ class BadgeClassAdmin(DjangoObjectActions, ModelAdmin):
         BadgeClassAlignmentInline,
         BadgeClassExtensionInline,
     ]
-    change_actions = ['redirect_issuer', 'redirect_instances', 'redirect_pathwaybadges']
+    change_actions = ['redirect_issuer', 'redirect_instances']
 
     def save_model(self, request, obj, form, change):
         force_resize = False
@@ -148,13 +148,6 @@ class BadgeClassAdmin(DjangoObjectActions, ModelAdmin):
         )
     redirect_issuer.label = "Issuer"
     redirect_issuer.short_description = "See this Issuer"
-
-    def redirect_pathwaybadges(self, request, obj):
-        return HttpResponseRedirect(
-            reverse('admin:pathway_pathwayelementbadge_changelist') + '?badgeclass__id={}'.format(obj.id)
-        )
-    redirect_pathwaybadges.label = "Pathway Badges"
-    redirect_pathwaybadges.short_description = "Pathway Badges"
 
 badgr_admin.register(BadgeClass, BadgeClassAdmin)
