+++ apps/issuer/models.py
@@ -318,14 +318,6 @@ class Issuer(ResizeUploadedImage,
     def cached_badgeclasses(self):
         return self.badgeclasses.all().order_by("created_at")
 
-    @cachemodel.cached_method(auto_publish=True)
-    def cached_pathways(self):
-        return self.pathway_set.filter(is_active=True)
-
-    @cachemodel.cached_method(auto_publish=True)
-    def cached_recipient_groups(self):
-        return self.recipientgroup_set.all()
-
     @property
     def image_preview(self):
         return self.image
@@ -504,12 +496,6 @@ class BadgeClass(ResizeUploadedImage,
                 models.Q(expires_at__isnull=True) | models.Q(expires_at__gt=timezone.now())).exists():
             raise ProtectedError("BadgeClass may only be deleted if all BadgeInstances have been revoked.", self)
 
-        if self.pathway_element_count() > 0:
-            raise ProtectedError("BadgeClass may only be deleted if all PathwayElementBadge have been removed.", self)
-
-        if len(self.cached_completion_elements()) > 0:
-            raise ProtectedError("Badge could not be deleted. It is being used as a pathway completion badge.", self)
-
         issuer = self.issuer
         super(BadgeClass, self).delete(*args, **kwargs)
         issuer.publish(publish_staff=False)
@@ -567,9 +553,6 @@ class BadgeClass(ResizeUploadedImage,
     @property
     def v1_api_recipient_count(self):
         return self.badgeinstances.filter(revoked=False).count()
-
-    def pathway_element_count(self):
-        return len(self.cached_pathway_elements())
 
     @cachemodel.cached_method(auto_publish=True)
     def cached_alignments(self):
@@ -642,14 +625,6 @@ class BadgeClass(ResizeUploadedImage,
 
     def get_extensions_manager(self):
         return self.badgeclassextension_set
-
-    @cachemodel.cached_method(auto_publish=True)
-    def cached_pathway_elements(self):
-        return [peb.element for peb in self.pathwayelementbadge_set.all()]
-
-    @cachemodel.cached_method(auto_publish=True)
-    def cached_completion_elements(self):
-        return [pce for pce in self.completion_elements.all()]
 
     def issue(self, recipient_id=None, evidence=None, narrative=None, notify=False, created_by=None, allow_uppercase=False, badgr_app=None, recipient_type=RECIPIENT_TYPE_EMAIL, **kwargs):
         return BadgeInstance.objects.create(
@@ -1045,16 +1020,6 @@ class BadgeInstance(BaseAuditedModel,
 
     def get_extensions_manager(self):
         return self.badgeinstanceextension_set
-
-    @property
-    def cached_recipient_profile(self):
-        from recipient.models import RecipientProfile
-        try:
-            return RecipientProfile.cached.get(recipient_identifier=self.recipient_identifier)
-        except RecipientProfile.MultipleObjectsReturned:
-            return RecipientProfile.objects.filter(recipient_identifier=self.recipient_identifier).first()
-        except RecipientProfile.DoesNotExist:
-            return None
 
     @property
     def recipient_user(self):
