--- apps/pathway/tests/__init__.py
@@ -1,1 +1,0 @@
-# Created by wiggins@concentricsky.com on 4/16/16.
