--- apps/recipient/__init__.py
@@ -1,1 +1,0 @@
-# Created by wiggins@concentricsky.com on 3/31/16.
